[Research] [th01/th02/th03] Document why graph_putsa_fx() uses row 0x29 for ANK
Last-minute research spurred by オップナー2608's research into a
rendering bug of the English patch for エーゲ海の雫 on a PC-9801DS/U2:

	https://github.com/Svipur/Aegean-Kai-No-Shizuku-PC98-English/issues/1

Though it seems as if ZUN's reason for using JIS row 0x29 is unrelated
to the KCG's code and dot access modes (since I can't reproduce these
differences anywhere) and can be entirely gathered from the PC-9801
Programmers' Bible, which we know for a fact that he was reading during
the development of TH01.

Part of P0323, funded by Ember2528.